+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.2.10...HEAD)
 ------------------------------------------------------------------------
 ### Added
-Adds `pyaud.BasePlugin` class for typing
+- Adds logger to `BasePlugin`
+- Adds `pyaud.BasePlugin` class for typing
 
 ### Fixed
 - Fixes returncode with tests